data.py moved into adapters
